requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
discord.py==2.2.3
doppler-env==0.3.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.29.0
SQLAlchemy==2.0.15
typing_extensions==4.6.2
urllib3==1.26.15
yarl==1.9.2
